Advance 64-bit support: fix remaining narrowing casts and fseek
- CDlgTagJumpList: fseek(SEEK_END) -> _fseeki64 for >2GB tags files
- my_strtok: int nLen/pnOffset -> ssize_t; update all callers
  (CGrepAgent, CGrepEnumKeys, CMacro, CType_Asm, CCommandLine)
- CColor_Quote::Match_QuoteStr: int nQuoteLen -> ssize_t
- CEditView_ExecCmd: int read_cntw -> DWORD, remove (int)sizeof casts
- AGENTS.md: mark P1/P2 tasks complete, add P4 manual test procedure

+++ sakura_core/_main/CCommandLine.cpp
@@ -246,7 +246,7 @@ void CCommandLine::ParseCommandLine( LPCWSTR pszCmdLineSrc, bool bResponse )
 	WCHAR	szPath[_MAX_PATH];
 	bool	bFind = false;				// ファイル名発見フラグ
 	bool	bParseOptDisabled = false;	// 2007.09.09 genta オプション解析を行わなず，ファイル名として扱う
-	int		nPos;
+	ssize_t	nPos;
 	int		i = 0;
 	if( pszCmdLineSrc[0] != L'-' ){
 		for( i = 0; i < int(std::size(szPath)); ++i ){
@@ -276,7 +276,7 @@ void CCommandLine::ParseCommandLine( LPCWSTR pszCmdLineSrc, bool bResponse )
 	CNativeW cmResponseFile = L"";
 	std::wstring strCmdLineWork{ pszCmdLineSrc };
 	LPWSTR pszCmdLineWork = std::data(strCmdLineWork);
-	int nCmdLineWorkLen = lstrlen( pszCmdLineWork );
+	ssize_t nCmdLineWorkLen = static_cast<ssize_t>( lstrlen( pszCmdLineWork ) );
 	LPWSTR pszToken = my_strtok<WCHAR>( pszCmdLineWork, nCmdLineWorkLen, &nPos, L" " );
 	while( pszToken != nullptr )
 	{
