+++ sakura_core/macro/CMacro.cpp
@@ -2329,14 +2329,14 @@ bool CMacro::HandleFunction(CEditView *View, EFunctionCode ID, VARIANT *Argument
 				if( !VariantToI4(varCopy, Arguments[0]) ) return false;
 				if( !VariantToBStr(varCopy2, Arguments[1]) ) return false;
 				std::vector<wchar_t> vStrMenu;
-				int nLen = (int)wcslen(varCopy2.Data.bstrVal);
+				ssize_t nLen = (ssize_t)wcslen(varCopy2.Data.bstrVal);
 				vStrMenu.assign( nLen + 1, L'\0' );
 				::wcsncpy_s(&vStrMenu[0], std::size(vStrMenu), varCopy2.Data.bstrVal, _TRUNCATE);
 				HMENU hMenu = ::CreatePopupMenu();
 				std::vector<HMENU> vHmenu;
 				vHmenu.push_back( hMenu );
 				HMENU hMenuCurrent = hMenu;
-				int nPos = 0;
+				ssize_t nPos = 0;
 				wchar_t* p;
 				int i = 1;
 				while ((p = my_strtok(&vStrMenu[0], nLen, &nPos, L","))) {
